feat: Consolidate settings page
